requirements.txt
pycryptodome==3.16.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
python-dotenv==0.21.0